pyproject.toml
[project]
name = "autosendmail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pre-commit>=4.5.1",
    "python-dotenv>=1.2.1",
]
